api with sanctum

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api\V2;
 
 use App\Models\Task;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
-
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,15 +16,9 @@ class TaskController extends Controller
      */
     public function index()
     {
-        return TaskResource::collection(auth()->user()->tasks()->get());
-    }
+        Gate::authorize('viewAny', Task::class);
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
+        return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
     /**
@@ -31,7 +26,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        Gate::authorize('create', Task::class);
+
         $task = $request->user()->tasks()->create($request->validated());
+
         return TaskResource::make($task);
     }
 
@@ -40,15 +38,9 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
-        return TaskResource::make($task);
-    }
+        Gate::authorize('view', $task);
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Task $task)
-    {
-        //
+        return TaskResource::make($task);
     }
 
     /**
@@ -56,15 +48,20 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        Gate::authorize('update', $task);
+
         $task->update($request->validated());
+
         return TaskResource::make($task);
     }
 
-    /*
+    /**
      * Remove the specified resource from storage.
      */
     public function destroy(Task $task)
     {
+        Gate::authorize('delete', $task);
+
         $task->delete();
 
         return response()->noContent();
